# Performance: Context holds SynchronizedSpan directly, not via HashMap