@jtescher 2023-09-13
Looks good, always nice to be able to do perf improvements without breaking api changes too 👍 what numbers are you seeing for the diff?