@shaun-cox 2023-09-13
> Looks good, always nice to be able to do perf improvements without breaking api changes too 👍 what numbers are you seeing for the diff?

From 70% to 170% speedup in checking context for active span, or is sampled or is recording.
Up to 58% reduction in conditional span creation time (140% speedup) when creating a child span only if currently recording.  Pasted results in the PR comment above.