@shaun-cox 2023-09-14
> What's the rationale for moving from per-module to per-crate imports? Is that converging on a common style or diverging from it?

I thought I was moving to the common style, as its what's offered by auto-complete/rust-analyzer when adding new imports.  I can revert if I didn't understand the current style tho...