@shaun-cox 2023-09-14
@djc, should be easier to review now.  (Sorry, I thought I had squashed my two commits previously, but didn't.)